Desktop/tonalt-bot/requirements.txt
python-telegram-bot==20.3
firebase-admin==6.2.0
python-dotenv==1.0.0
flask==3.0.0
